Move BVHNode from hitable.rs to bvhnode.rs
